requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-data-tables==12.4.3
azure-storage-blob==12.17.0
PyYAML==6.0.1
tqdm==2.2.3
pandas==2.0.3
azure-storage-queue==12.6.0

